Implemented CB-code for Entities (part 3, final)

+++ src/main/java/net/minecraft/entity/passive/EntityRabbit.java
@@ -68,6 +68,11 @@ public class EntityRabbit extends EntityAnimal
         this.setSize(0.4F, 0.5F);
         this.jumpHelper = new RabbitJumpHelper(this);
         this.moveHelper = new RabbitMoveHelper(this);
+        this.initializePathFinderGoals(); // CraftBukkit - moved code
+    }
+
+    // CraftBukkit start - code from constructor
+    public void initializePathFinderGoals(){
         this.setMovementSpeed(0.0D);
     }
 
@@ -587,11 +592,21 @@ public class EntityRabbit extends EntityAnimal
 
                         if (integer.intValue() == 0)
                         {
+                            if (org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.rabbit, blockpos, Blocks.AIR, 0).isCancelled()) {
+                                return;
+                            }
                             world.setBlockState(blockpos, Blocks.AIR.getDefaultState(), 2);
                             world.destroyBlock(blockpos, true);
                         }
                         else
                         {
+                            if (org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(
+                                    this.rabbit,
+                                    blockpos,
+                                    block, block.getMetaFromState(iblockstate.withProperty(BlockCarrot.AGE, Integer.valueOf(integer.intValue() - 1)))
+                            ).isCancelled()) {
+                                return;
+                            }
                             world.setBlockState(blockpos, iblockstate.withProperty(BlockCarrot.AGE, Integer.valueOf(integer.intValue() - 1)), 2);
                             world.playEvent(2001, blockpos, Block.getStateId(iblockstate));
                         }
